+++ src/main/java/org/zerock/controller/LoginController.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.zerock.domain.MemberVO;
 import org.zerock.service.LoginService;
 
@@ -21,7 +23,8 @@ import lombok.extern.log4j.Log4j2;
 @Controller
 @Log4j2
 public class LoginController {
-
+	private LoginService service;
+	
 	@Setter(onMethod_ = @Autowired )
 	private LoginService loginService ; //로그인을 위한 매퍼 서비스 객체
 	
@@ -52,5 +55,23 @@ public class LoginController {
 		
 		return "/findPw";
 	}
+	
+	@PostMapping("/join")
+	@PreAuthorize("isAuthenticated()")
+	public String join(MemberVO member, RedirectAttributes rttr) {
+
+		log.info("==========================");
+
+		log.info("join: " + member);
+
+		log.info("==========================");
+
+		service.join(member);
+
+		rttr.addFlashAttribute("result");
+		
+		return "redirect:/join";
+	}
+	
 	
 }
