+++ src/main/java/org/zerock/service/LoginService.java
@@ -13,4 +13,6 @@ public interface LoginService {
 	public MemberVO findPw (String email ,String id); // id와 이메일로 패스워드를 찾는 메서드 
 	
 	public String resetPassword(String newpassword, String id); //비밀번호를 초기화 하는메서드(implement에서 바로 sql실행예정)
+	
+	public int join(MemberVO member); //회원가입용
 }
