--- src/main/java/org/zerock/service/MemberService.java
@@ -1,7 +1,0 @@
-package org.zerock.service;
-
-import org.zerock.domain.MemberVO;
-
-public interface MemberService {
-	public int join(MemberVO member);
-}
