+++ src/test/java/org/zerock/mapper/LoginMapperTests.java
@@ -20,10 +20,10 @@ import lombok.extern.log4j.Log4j2;
 	  "file:src/main/webapp/WEB-INF/spring/security-context.xml"
 	  })
 @Log4j2
-public class MemberMapperTests {
+public class LoginMapperTests {
 	
 	@Setter(onMethod_ = @Autowired)
-	private MemberMapper mapper;
+	private LoginMapper mapper;
 	
 	@Setter(onMethod_ = @Autowired)
 	private PasswordEncoder pwencoder;
